+++ AddOns/Details/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Details! Damage Meter
 
-## [Details.20251105.13857.166](https://github.com/Tercioo/Details-Damage-Meter/tree/Details.20251105.13857.166) (2025-11-05)
-[Full Changelog](https://github.com/Tercioo/Details-Damage-Meter/compare/Details.20251023.13811.166...Details.20251105.13857.166) 
+## [Details.20251212.13911.166](https://github.com/Tercioo/Details-Damage-Meter/tree/Details.20251212.13911.166) (2025-12-12)
+[Full Changelog](https://github.com/Tercioo/Details-Damage-Meter/compare/Details.20251105.13857.166...Details.20251212.13911.166) 
 
-- Several fixes for Vanilla wow.  
+- Tag as release  
+- Bug fixes, framework update, subtle changes for Midnight.  
+- Merge pull request #978 from Hollicsh/master  
+    Update the .toc files  
+- Update the .toc files  
+- Added icon tags for the encounter journal dump.  
+- Dump journal difficulty selectable.  
+- Encounter Journal Dump  
+- Fixes and add /details bossspells <boss name>  to dump boss spells